Coerce `document_id` to integer
Document ids can be strings with additional slug elements
which the DocumentCloud Rails app API coerces to an integer to find the
actual record.

DocumentCloud does so here:

https://github.com/documentcloud/documentcloud/blob/2dc6534ac9e2eea7135497d0135a6dc50e4252e4/app/controllers/api_controller.rb#L257

The feature of the Ruby language it relies on is that any string
starting with a number will be truncated at the first non-numerical
character.

```ruby
'123-abc'.to_i # => 123
```

Why is this an issue?

Document id slugs can include other UTF8 characters. I ran into this
issue with Document ids like:

'2713540-Ōrākei-Local-Board-20151210'

due to the macrons with Māori place names.

Which generated an `URI::InvalidURIError: URI must be ascii only`
exception.